+++ modules/renderer/src/com/genir/renderer/bridge/GL15.java
@@ -26,7 +26,7 @@ public class GL15 {
     public static void glDeleteBuffers(int buffer) {
         record glDeleteBuffers(int buffer) implements GLCommand {
             @Override
-            public void run(Context context, int[] args, int offset) {
+            public void run(Context context, float[] args, int offset) {
                 org.lwjgl.opengl.GL15.glDeleteBuffers(buffer);
             }
         }
@@ -37,7 +37,7 @@ public class GL15 {
     public static void glDeleteBuffers(IntBuffer buffers) {
         record glDeleteBuffers(IntBuffer buffers) implements GLCommand {
             @Override
-            public void run(Context context, int[] args, int offset) {
+            public void run(Context context, float[] args, int offset) {
                 org.lwjgl.opengl.GL15.glDeleteBuffers(buffers);
             }
         }
@@ -49,7 +49,7 @@ public class GL15 {
     public static void glBindBuffer(int target, int buffer) {
         record glBindBuffer(int target, int buffer) implements GLCommand {
             @Override
-            public void run(Context context, int[] args, int offset) {
+            public void run(Context context, float[] args, int offset) {
                 org.lwjgl.opengl.GL15.glBindBuffer(target, buffer);
             }
         }
@@ -63,7 +63,7 @@ public class GL15 {
     public static void glBufferData(int target, long data_size, int usage) {
         record glBufferData(int target, long data_size, int usage) implements GLCommand {
             @Override
-            public void run(Context context, int[] args, int offset) {
+            public void run(Context context, float[] args, int offset) {
                 org.lwjgl.opengl.GL15.glBufferData(target, data_size, usage);
             }
         }
@@ -74,7 +74,7 @@ public class GL15 {
     public static void glBufferData(int target, FloatBuffer data, int usage) {
         record glBufferData(int target, FloatBuffer data, int usage) implements GLCommand {
             @Override
-            public void run(Context context, int[] args, int offset) {
+            public void run(Context context, float[] args, int offset) {
                 org.lwjgl.opengl.GL15.glBufferData(target, data, usage);
             }
         }
@@ -86,7 +86,7 @@ public class GL15 {
     public static void glBufferData(int target, ByteBuffer data, int usage) {
         record glBufferData(int target, ByteBuffer data, int usage) implements GLCommand {
             @Override
-            public void run(Context context, int[] args, int offset) {
+            public void run(Context context, float[] args, int offset) {
                 org.lwjgl.opengl.GL15.glBufferData(target, data, usage);
             }
         }
@@ -98,7 +98,7 @@ public class GL15 {
     public static void glBufferData(int target, ShortBuffer data, int usage) {
         record glBufferData(int target, ShortBuffer data, int usage) implements GLCommand {
             @Override
-            public void run(Context context, int[] args, int offset) {
+            public void run(Context context, float[] args, int offset) {
                 org.lwjgl.opengl.GL15.glBufferData(target, data, usage);
             }
         }
@@ -110,7 +110,7 @@ public class GL15 {
     public static void glBufferData(int target, IntBuffer data, int usage) {
         record glBufferData(int target, IntBuffer data, int usage) implements GLCommand {
             @Override
-            public void run(Context context, int[] args, int offset) {
+            public void run(Context context, float[] args, int offset) {
                 org.lwjgl.opengl.GL15.glBufferData(target, data, usage);
             }
         }
@@ -122,7 +122,7 @@ public class GL15 {
     public static void glBufferSubData(int target, long offset, FloatBuffer data) {
         record glBufferSubData(int target, long offset, FloatBuffer data) implements GLCommand {
             @Override
-            public void run(Context context, int[] args, int offset) {
+            public void run(Context context, float[] args, int offset) {
                 org.lwjgl.opengl.GL15.glBufferSubData(target, offset, data);
             }
         }
@@ -134,7 +134,7 @@ public class GL15 {
     public static void glBufferSubData(int target, long offset, ShortBuffer data) {
         record glBufferSubData(int target, long offset, ShortBuffer data) implements GLCommand {
             @Override
-            public void run(Context context, int[] args, int offset) {
+            public void run(Context context, float[] args, int offset) {
                 org.lwjgl.opengl.GL15.glBufferSubData(target, offset, data);
             }
         }
@@ -146,7 +146,7 @@ public class GL15 {
     public static void glBufferSubData(int target, long offset, IntBuffer data) {
         record glBufferSubData(int target, long offset, IntBuffer data) implements GLCommand {
             @Override
-            public void run(Context context, int[] args, int offset) {
+            public void run(Context context, float[] args, int offset) {
                 org.lwjgl.opengl.GL15.glBufferSubData(target, offset, data);
             }
         }
@@ -158,7 +158,7 @@ public class GL15 {
     public static void glBufferSubData(int target, long offset, ByteBuffer data) {
         record glBufferSubData(int target, long offset, ByteBuffer data) implements GLCommand {
             @Override
-            public void run(Context context, int[] args, int offset) {
+            public void run(Context context, float[] args, int offset) {
                 org.lwjgl.opengl.GL15.glBufferSubData(target, offset, data);
             }
         }
