+++ modules/renderer/src/com/genir/renderer/bridge/GL43.java
@@ -25,7 +25,7 @@ public class GL43 {
     public static void glShaderStorageBlockBinding(int program, int storageBlockIndex, int storageBlockBinding) {
         record glShaderStorageBlockBinding(int program, int storageBlockIndex, int storageBlockBinding) implements GLCommand {
             @Override
-            public void run(Context context, int[] args, int offset) {
+            public void run(Context context, float[] args, int offset) {
                 org.lwjgl.opengl.GL43.glShaderStorageBlockBinding(program, storageBlockIndex, storageBlockBinding);
             }
         }
@@ -36,7 +36,7 @@ public class GL43 {
     public static void glDispatchCompute(int num_groups_x, int num_groups_y, int num_groups_z) {
         record glDispatchCompute(int num_groups_x, int num_groups_y, int num_groups_z) implements GLCommand {
             @Override
-            public void run(Context context, int[] args, int offset) {
+            public void run(Context context, float[] args, int offset) {
                 org.lwjgl.opengl.GL43.glDispatchCompute(num_groups_x, num_groups_y, num_groups_z);
             }
         }
@@ -47,7 +47,7 @@ public class GL43 {
     public static void glDebugMessageCallback(KHRDebugCallback callback) {
         record glDebugMessageCallback(KHRDebugCallback callback) implements GLCommand {
             @Override
-            public void run(Context context, int[] args, int offset) {
+            public void run(Context context, float[] args, int offset) {
                 org.lwjgl.opengl.GL43.glDebugMessageCallback(callback);
             }
         }
@@ -58,7 +58,7 @@ public class GL43 {
     public static void glDebugMessageControl(int source, int type, int severity, IntBuffer ids, boolean enabled) {
         record glDebugMessageControl(int source, int type, int severity, IntBuffer ids, boolean enabled) implements GLCommand {
             @Override
-            public void run(Context context, int[] args, int offset) {
+            public void run(Context context, float[] args, int offset) {
                 org.lwjgl.opengl.GL43.glDebugMessageControl(source, type, severity, ids, enabled);
             }
         }
@@ -70,7 +70,7 @@ public class GL43 {
     public static void glInvalidateBufferData(int buffer) {
         record glInvalidateBufferData(int buffer) implements GLCommand {
             @Override
-            public void run(Context context, int[] args, int offset) {
+            public void run(Context context, float[] args, int offset) {
                 org.lwjgl.opengl.GL43.glInvalidateBufferData(buffer);
             }
         }
