+++ modules/renderer/src/com/genir/renderer/bridge/context/ListManager.java
@@ -32,10 +32,10 @@ public class ListManager {
         return mode != 0;
     }
 
-    public void record(GLCommand command, int[] args, int argsOffset) {
+    public void record(GLCommand command, float[] args, int argsOffset) {
         newList.add(command);
 
-        int argsSize = args[argsOffset];
+        int argsSize = (int) args[argsOffset];
         for (int j = argsOffset; j < argsOffset + argsSize; j++) {
             newList.args[newList.argsOffset++] = args[j];
         }
@@ -72,14 +72,13 @@ public class ListManager {
 
         Frame listToCall = lists.get(list);
         if (listToCall != null) {
-            int[] args = listToCall.args;
+            float[] args = listToCall.args;
             int argsOffset = 0;
 
             // for-each loop over a list is a performance bottleneck, according to a profiler.
             // Simple for loop over an array is much faster.
             for (int i = 0; i < listToCall.commandsSize; i++) {
-
-                int argsSize = args[argsOffset];
+                int argsSize = (int) args[argsOffset];
                 listToCall.commands[i].run(context, args, argsOffset);
                 argsOffset += argsSize;
             }
