+++ modules/renderer/src/com/genir/renderer/bridge/context/commands/GLCommand.java
@@ -8,5 +8,5 @@ import com.genir.renderer.bridge.context.Context;
 public interface GLCommand {
     // Execute the GL command. int[] args is used by
     // memory-optimized commands to avoid heap allocation.
-    void run(Context context, int[] args, int offset);
+    void run(Context context, float[] args, int offset);
 }
